+++ test/runner_normalise_doctest1_test.exs
@@ -1,0 +1,10 @@
+defmodule HarnaisExampleDoctestRunnerNormalise1Test do
+
+  use HarnaisHelpersTest
+
+  import Harnais.Runner.Normalise
+
+  doctest Harnais.Runner.Normalise
+
+end
+