+++ test/test_helper.exs
@@ -7,6 +7,7 @@ defmodule HarnaisHelpersTest do
     quote do
       use ExUnit.Case, async: true
       use Harnais
+      alias Harnais.Runner.Normalise, as: HRN
     end
 
   end